sdfec_cmodel_python: Add xoroshiro128 implementing with fowrarding
This patch introduces an implementation of the xoroshiro128plus [0] random
number generator which has been extended with functions that allow
forwarding to an arbitrary point in O(log2(N)) time. While there are
more storage efficient algorithms available [1], the performance of this
approach is sufficient for my purposes (Forwarding by billions of steps
takes on the order of 10 µs from python).

[0] https://arxiv.org/pdf/1805.01407.pdf
[1] http://www.math.sci.hiroshima-u.ac.jp/m-mat/MT/ARTICLES/jumpf2-printed.pdf

Signed-off-by: David Winter <dastw@gmx.net>