sdfec_cmodel_python: Add LDPC BER tester model
This commit introduces a simple model of the LDPC BER Tester as it is
available in hardware. This model allows the user to simulate specific
blocks in software, which have been deemed interesting by the hardware
simulation, and inspect both the inputs and outputs in various
configurations through the python interface.

Signed-off-by: David Winter <dastw@gmx.net>